+++ src/db/repositories/user-repository.ts
@@ -2,7 +2,7 @@ import { db } from '@/db'
 import { schema } from '@/db/schema'
 import type { InsertUserModel } from '@/domain/entities/user'
 import type { UpdateUserInput } from '@/domain/services/users/update-user'
-import { eq, sql } from 'drizzle-orm'
+import { desc, eq, like, sql } from 'drizzle-orm'
 
 export async function getUserByEmail(email: string) {
   return db
@@ -93,4 +93,28 @@ export async function getProUsersDetails() {
     )
     .leftJoin(schema.userItems, eq(schema.users.id, schema.userItems.userId))
     .groupBy(schema.users.id, schema.userPreferences.id)
+}
+
+export async function listUsersByUsernameLike(username: string) {
+  return db
+    .select({
+      id: schema.users.id,
+      username: schema.users.username,
+      avatarUrl: schema.users.avatarUrl,
+      isFollowed: sql<boolean>`EXISTS (
+        SELECT 1 FROM ${schema.followers}
+        WHERE ${schema.followers.followedId} = ${schema.users.id}
+      )`,
+    })
+    .from(schema.users)
+    .where(like(schema.users.username, `%${username}%`))
+    .orderBy(
+      desc(
+        sql`CASE WHEN EXISTS (
+          SELECT 1 FROM ${schema.followers}
+          WHERE ${schema.followers.followedId} = ${schema.users.id}
+        ) THEN 1 ELSE 0 END`
+      )
+    )
+    .limit(10)
 }