+++ src/http/schemas/users.ts
@@ -122,3 +122,20 @@ export const getUserPreferencesResponseSchema = {
     userPreferences: createSelectSchema(schema.userPreferences).nullable(),
   }),
 }
+
+export const searchUsersByUsernameQuerySchema = z.object({
+  username: z.string(),
+})
+
+export const searchUsersByUsernameResponseSchema = {
+  200: z.object({
+    users: z.array(
+      z.object({
+        id: z.string(),
+        username: z.string(),
+        avatarUrl: z.string().nullable(),
+        isFollowed: z.boolean(),
+      })
+    ),
+  }),
+}