+++ CHANGELOG.md
@@ -7,6 +7,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Add recommendations cron job [(#105)](https://github.com/plotwist-app/plotwist-backend/pull/105)
 
+## 1.12.1
+
+### Fixed
+
+- Fix search users [(#106)](https://github.com/plotwist-app/plotwist-backend/pull/106)
+
+## 1.12.0
+
+### Added
+
+- Add search users by username endpoint [(#104)](https://github.com/plotwist-app/plotwist-backend/pull/104)
+
 ## 1.11.1
 
 ### Fixed